Find several possible distributions.

This function aims to find several possible distribution that would give rise to the observed sample parameters. For that, you need to pass a list of parameters, created with set_parameters

find_possible_distributions(
  parameters,
  n_distributions = 10,
  seed = NULL,
  return_tibble = TRUE,
  return_failures = FALSE
)

Arguments

parameters

List of parameters, see set_parameters

n_distributions

The target number of distributions to return.

seed

An integer to use as the seed for random number generation. Set this in scripts to ensure reproducibility.

return_tibble

Should a tibble, rather than a list, be returned? Requires the tibble-package, ignored if that package is not available.

return_failures

Should distributions that failed to produce the desired SD be returned? Defaults to false

Value

A tibble or list (depending on the return_tibble argument) with:

outcome

success or failure - character

distribution

The distribution that was found (if success) / that had the closest variance (if failure) - numeric

mean

The exact mean of the distribution - numeric

sd

The SD of the distribution that was found (success) / that came closest (failure) - numeric

iterations

The number of iterations required to achieve the specified SD - numeric - the first time this distribution was found

Examples

sprite_parameters <- set_parameters(mean = 2.2, sd = 1.3, n_obs = 20,
                                    min_val = 1, max_val = 5)

find_possible_distributions(sprite_parameters, 5, seed = 1234)
#> # A tibble: 5 × 6
#>      id outcome distribution  mean    sd iterations
#>   <int> <chr>   <list>       <dbl> <dbl>      <dbl>
#> 1     1 success <dbl [20]>    2.2   1.32          3
#> 2     2 success <dbl [20]>    2.25  1.25          1
#> 3     3 success <dbl [20]>    2.2   1.32          6
#> 4     4 success <dbl [20]>    2.25  1.33          1
#> 5     5 success <dbl [20]>    2.25  1.33          2